Adding sparse weight support
